// frontend/src/app/models/takmicenje.ts
export class Takmicenje{
    _id: string;
    ime: string;
    format: string;
    pol: string;
    sport: string;
    disciplina: string;
    pocetak: string;
    kraj: string;
    lokacija: string;
    delegat: string;
    sportisti: string[];
    prvoMesto: string;
    drugoMesto: string;
    treceMesto: string;
    formirano: boolean;
    zavrseno: boolean;
}